fix(stars): reconcile monitoring and finalize refunds
