--- .github/workflows/pr310-review-round-fixes.yml
@@ -1,103 +1,0 @@
-name: Validate PR 310 review-round fixes
-
-on:
-  pull_request:
-    branches:
-      - main
-    types:
-      - opened
-      - reopened
-      - synchronize
-
-permissions:
-  contents: write
-
-concurrency:
-  group: pr310-review-round-fixes-v6
-  cancel-in-progress: true
-
-jobs:
-  patch-test-commit:
-    if: github.event.pull_request.number == 310 && github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    timeout-minutes: 25
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: hotfix/stars-payment-safety
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: npm
-
-      - name: Prepare deterministic patcher
-        run: |
-          python3 - <<'PY'
-          from pathlib import Path
-          path = Path('scripts/pr310-review-round-fixes.py')
-          text = path.read_text()
-          old_engine = 'updated, count = re.subn(pattern, replacement, text, count=1, flags=flags)'
-          new_engine = 'updated, count = re.subn(pattern, lambda _match: replacement, text, count=1, flags=flags)'
-          if old_engine not in text:
-              raise SystemExit('regex replacement engine marker not found')
-          text = text.replace(old_engine, new_engine, 1)
-
-          old_fixture = """    payMonitorBundle('expired-week', 'renewed-user', 'expired-week-charge', 199, expiredWeekPaidAt);\n\n    insertMonitorBundle('later-month', 'renewed-user', 'monitor_month', 499, now, 5);"""
-          new_fixture = """    payMonitorBundle('expired-week', 'renewed-user', 'expired-week-charge', 199, expiredWeekPaidAt);\n    db.prepare(\n      \"UPDATE star_monitor_entitlements SET expires_at = ? WHERE user_id = 'renewed-user'\",\n    ).run(expiredWeekPaidAt + (7 * 24 * 60 * 60));\n\n    insertMonitorBundle('later-month', 'renewed-user', 'monitor_month', 499, now, 5);"""
-          if old_fixture not in text:
-              raise SystemExit('expired-grant fixture marker not found')
-          path.write_text(text.replace(old_fixture, new_fixture, 1))
-          PY
-
-      - name: Apply review-round fixes
-        shell: bash
-        run: |
-          set -o pipefail
-          python3 scripts/pr310-review-round-fixes.py 2>&1 | tee patch.log
-
-      - name: Upload patch diagnostics
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: pr310-patch-log-${{ github.run_id }}
-          path: patch.log
-          if-no-files-found: warn
-
-      - name: Prepare runtime data directory
-        run: |
-          sudo mkdir -p /data
-          sudo chmod 777 /data
-
-      - name: Install dependencies
-        run: npm ci --legacy-peer-deps
-
-      - name: Build TypeScript
-        run: npm run build
-
-      - name: Run full Jest suite
-        shell: bash
-        run: |
-          set -o pipefail
-          timeout 360s npm test -- --runInBand --forceExit 2>&1 | tee jest.log
-
-      - name: Upload Jest diagnostics
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: pr310-jest-log-${{ github.run_id }}
-          path: jest.log
-          if-no-files-found: warn
-
-      - name: Commit validated fixes
-        shell: bash
-        run: |
-          rm -f .github/workflows/pr310-review-round-fixes.yml
-          rm -f scripts/pr310-review-round-fixes.py
-          rm -f patch.log jest.log
-          git config user.name "chatgpt-codex-connector[bot]"
-          git config user.email "199175422+chatgpt-codex-connector[bot]@users.noreply.github.com"
-          git add -A
-          git commit -m "fix(stars): reconcile monitoring and finalize refunds"
-          git push origin HEAD:hotfix/stars-payment-safety
